Add reminder for the final RAL tech day
This is going to be the last tech day of the year. While many schools
have likely broken up for Easter, I'm hoping that some of them will
still see this.

Given our previous experience with tech days, it seems that sending
a reminder email at about this duration before the tech day is useful
in ensuring a good turnout.

Patch set 2: Move send date forwards

Change-Id: If4afd3419e2533b2758e83db665bec593f264415