Add email variants offering mentoring to all teams
These are all intended to go out at the same time, however they
adapt the message a bit to the recipients. This aims to encourage
the less active teams to get a mentor more than the more active ones,
while ensuring that everyone is aware that mentoring is available.

As such, these are largely a copy/paste of the messages sent in
Discord (see https://github.com/srobo/team-emails/pull/76), with
light adaptation for being sent by email to the team-leaders rather
than direct to the teams.

This also includes brief messaging about the rules/simulator updates
being available, though that isn't the focus of these emails.