Add final competition details email
This is probably going to be the last email we bulk send to teams.
Normally we'd send this a bit closer to the event, however given
the timing of Easter and the related holidays we need to send this
a little earlier to hopefully give people time to see it when
they're not on holiday and before the event.