Merge information about guests and public access
The event being open to the public is part of our approach to media
consent, so even if we had exclusive access to the venue we'd still
be letting members of the public in. Therefore, whether we have
exclusive access to the venue isn't especially important to mention
in this email to teams. We do however want to tell them that it's
public, so folding this into the other information about guests
makes sense.