Add a check-in email I'd like to send to all teams
Partly this is so that we gather check-in info from all the teams,
but also provides a place to ask why they're not in Discord.

I've deliberately kept this short and personliased so it comes
across more as a direct email than a mailshot. I'll still send it
from teams@.